完成了设备自动轮询、设备数据缓存。设备日志储存，设备日志常用字段的缓存。策略获取json。

+++ SES-pojo/src/main/java/com/SES/entity/DeviceLog.java
@@ -40,5 +40,5 @@ public class DeviceLog implements Serializable {
 
     private Float power; // 单位：W
 
-    private Integer energyConsumption; // 单位：kWh
+    private Float energyConsumption; // 单位：kWh
 }
